Issue-264: Moved BaseVH and BaseRecyclerViewAdapter.kt to new packages.
