+++ app/src/main/java/com/kirchhoff/movies/ui/screens/core/movies/adapter/MovieVH.kt
@@ -6,7 +6,7 @@ import android.widget.TextView
 import com.kirchhoff.movies.R
 import com.kirchhoff.movies.core.extensions.downloadPoster
 import com.kirchhoff.movies.core.extensions.inflate
-import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
+import com.kirchhoff.movies.core.ui.recyclerview.adapter.viewholder.BaseVH
 import com.kirchhoff.movies.data.ui.main.UIMovie
 
 class MovieVH(parent: ViewGroup) : BaseVH<UIMovie>(parent.inflate(R.layout.item_movie)) {
