+++ app/src/main/java/com/kirchhoff/movies/ui/screens/core/movies/adapter/MoviesListAdapter.kt
@@ -1,8 +1,8 @@
 package com.kirchhoff.movies.ui.screens.core.movies.adapter
 
 import android.view.ViewGroup
-import com.kirchhoff.movies.core.ui.recyclerview.BaseRecyclerViewAdapter
-import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
+import com.kirchhoff.movies.core.ui.recyclerview.adapter.BaseRecyclerViewAdapter
+import com.kirchhoff.movies.core.ui.recyclerview.adapter.viewholder.BaseVH
 import com.kirchhoff.movies.data.ui.main.UIMovie
 
 class MoviesListAdapter(clickListener: OnItemClickListener<UIMovie>) :
