+++ app/src/main/java/com/kirchhoff/movies/ui/screens/details/movie/adapters/TrailersListAdapter.kt
@@ -1,8 +1,8 @@
 package com.kirchhoff.movies.ui.screens.details.movie.adapters
 
 import android.view.ViewGroup
-import com.kirchhoff.movies.core.ui.recyclerview.BaseRecyclerViewAdapter
-import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
+import com.kirchhoff.movies.core.ui.recyclerview.adapter.BaseRecyclerViewAdapter
+import com.kirchhoff.movies.core.ui.recyclerview.adapter.viewholder.BaseVH
 import com.kirchhoff.movies.data.ui.details.movie.UITrailer
 
 class TrailersListAdapter(trailersList: List<UITrailer>, clickListener: OnItemClickListener<UITrailer>) :
