+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/movies/MoviesFragment.kt
@@ -2,7 +2,7 @@ package com.kirchhoff.movies.ui.screens.main.movies
 
 import com.kirchhoff.movies.R
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
-import com.kirchhoff.movies.core.ui.recyclerview.BaseRecyclerViewAdapter
+import com.kirchhoff.movies.core.ui.recyclerview.adapter.BaseRecyclerViewAdapter
 import com.kirchhoff.movies.data.ui.main.UIMovie
 import com.kirchhoff.movies.ui.screens.core.PaginatedScreenFragment
 import com.kirchhoff.movies.ui.screens.core.movies.adapter.MoviesListAdapter
