+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/persons/adapter/PersonVH.kt
@@ -5,7 +5,7 @@ import android.widget.ImageView
 import com.kirchhoff.movies.R
 import com.kirchhoff.movies.core.extensions.downloadAvatar
 import com.kirchhoff.movies.core.extensions.inflate
-import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
+import com.kirchhoff.movies.core.ui.recyclerview.adapter.viewholder.BaseVH
 import com.kirchhoff.movies.data.ui.main.UIPerson
 
 class PersonVH(parent: ViewGroup) : BaseVH<UIPerson>(parent.inflate(R.layout.item_person)) {
