+++ app/src/main/java/com/kirchhoff/movies/ui/screens/similar/tv/SimilarTvsFragment.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import android.view.View
 import com.kirchhoff.movies.R
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
-import com.kirchhoff.movies.core.ui.recyclerview.BaseRecyclerViewAdapter
+import com.kirchhoff.movies.core.ui.recyclerview.adapter.BaseRecyclerViewAdapter
 import com.kirchhoff.movies.data.ui.main.UITv
 import com.kirchhoff.movies.ui.screens.core.PaginatedScreenFragment
 import com.kirchhoff.movies.ui.screens.core.tvs.adapter.TvsListAdapter
