+++ creditsview/src/main/java/com/kirchhoff/movies/creditsview/adapter/CreditVH.kt
@@ -5,7 +5,7 @@ import android.widget.ImageView
 import android.widget.TextView
 import com.kirchhoff.movies.core.extensions.downloadPoster
 import com.kirchhoff.movies.core.extensions.inflate
-import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
+import com.kirchhoff.movies.core.ui.recyclerview.adapter.viewholder.BaseVH
 import com.kirchhoff.movies.creditsview.CreditsView
 import com.kirchhoff.movies.creditsview.R
