+++ creditsview/src/main/java/com/kirchhoff/movies/creditsview/adapter/CreditsAdapter.kt
@@ -1,8 +1,8 @@
 package com.kirchhoff.movies.creditsview.adapter
 
 import android.view.ViewGroup
-import com.kirchhoff.movies.core.ui.recyclerview.BaseRecyclerViewAdapter
-import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
+import com.kirchhoff.movies.core.ui.recyclerview.adapter.BaseRecyclerViewAdapter
+import com.kirchhoff.movies.core.ui.recyclerview.adapter.viewholder.BaseVH
 import com.kirchhoff.movies.creditsview.CreditsView
 
 class CreditsAdapter(itemClickListener: OnItemClickListener<CreditsView.CreditsInfo>) :
