+++ keywordsview/src/main/java/com/kirchhoff/movies/keywordsview/adapter/KeywordVH.kt
@@ -3,7 +3,7 @@ package com.kirchhoff.movies.keywordsview.adapter
 import android.view.ViewGroup
 import android.widget.TextView
 import com.kirchhoff.movies.core.extensions.inflate
-import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
+import com.kirchhoff.movies.core.ui.recyclerview.adapter.viewholder.BaseVH
 import com.kirchhoff.movies.keywordsview.R
 
 class KeywordVH(parent: ViewGroup) : BaseVH<String>(parent.inflate(R.layout.item_keyword)) {
